[#32]refactor: classroomAccount -> classroom.account
