+++ src/main/java/soma/edupimeta/classroom/account/ClassroomAccountOpenApi.java
@@ -1,4 +1,4 @@
-package soma.edupimeta.classroomAccount;
+package soma.edupimeta.classroom.account;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -7,8 +7,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
-import soma.edupimeta.classroomAccount.models.CreateClassroomAccountRequest;
-import soma.edupimeta.classroomAccount.service.domain.ClassroomAccount;
+import soma.edupimeta.classroom.account.models.CreateClassroomAccountRequest;
+import soma.edupimeta.classroom.account.service.domain.ClassroomAccount;
 
 @Tag(name = "ClassroomAccount", description = "ClassroomAccount API")
 public interface ClassroomAccountOpenApi {
@@ -18,7 +18,7 @@ public interface ClassroomAccountOpenApi {
         @ApiResponse(responseCode = "200", description = "게스트 등록에 성공했습니다.", content = @Content(mediaType = "application/json")),
         @ApiResponse(responseCode = "400", description = "이미 등록된 계정입니다.", content = @Content(mediaType = "application/json")),
     })
-    ResponseEntity<ClassroomAccount> registerGuest(
+    ResponseEntity<ClassroomAccount> registerClassroomAccount(
         @RequestBody CreateClassroomAccountRequest createClassroomAccountRequest
     );
 }
