+++ src/main/java/soma/edupimeta/classroom/account/service/ClassroomAccountService.java
@@ -1,15 +1,15 @@
-package soma.edupimeta.classroomAccount.service;
+package soma.edupimeta.classroom.account.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import soma.edupimeta.classroom.account.exception.ClassroomAccountErrorEnum;
+import soma.edupimeta.classroom.account.exception.ClassroomAccountException;
+import soma.edupimeta.classroom.account.service.domain.ClassroomAccount;
+import soma.edupimeta.classroom.account.service.domain.ClassroomAccountRole;
+import soma.edupimeta.classroom.account.service.repository.ClassroomAccountRepository;
 import soma.edupimeta.classroom.exception.ClassroomErrorEnum;
 import soma.edupimeta.classroom.exception.ClassroomException;
 import soma.edupimeta.classroom.service.repository.ClassroomRepository;
-import soma.edupimeta.classroomAccount.exception.ClassroomAccountErrorEnum;
-import soma.edupimeta.classroomAccount.exception.ClassroomAccountException;
-import soma.edupimeta.classroomAccount.service.domain.ClassroomAccount;
-import soma.edupimeta.classroomAccount.service.domain.ClassroomAccountRole;
-import soma.edupimeta.classroomAccount.service.repository.ClassroomAccountRepository;
 
 @Service
 @RequiredArgsConstructor
@@ -18,7 +18,7 @@ public class ClassroomAccountService {
     private final ClassroomAccountRepository classroomAccountRepository;
     private final ClassroomRepository classroomRepository;
 
-    public ClassroomAccount registerGuest(Long accountId, Long classroomId) {
+    public ClassroomAccount registerClassroomAccount(Long accountId, Long classroomId, ClassroomAccountRole role) {
         if (isExistsClassroom(classroomId)) {
             throw new ClassroomException(ClassroomErrorEnum.CLASSROOM_NOT_FOUND);
         }
@@ -30,7 +30,7 @@ public class ClassroomAccountService {
         ClassroomAccount classroomAccount = ClassroomAccount.builder()
             .accountId(accountId)
             .classroomId(classroomId)
-            .role(ClassroomAccountRole.HOST)
+            .role(role)
             .build();
 
         return addClassroomAccount(classroomAccount);
@@ -41,7 +41,7 @@ public class ClassroomAccountService {
     }
 
     private boolean isDuplicate(Long accountId, Long classroomId) {
-        return !classroomAccountRepository.existsByAccountIdAndClassroomId(accountId, classroomId);
+        return classroomAccountRepository.existsByAccountIdAndClassroomId(accountId, classroomId);
     }
 
     private ClassroomAccount addClassroomAccount(ClassroomAccount classroomAccount) {
