+++ src/main/java/soma/edupimeta/classroom/account/service/domain/ClassroomAccountRole.java
@@ -1,0 +1,6 @@
+package soma.edupimeta.classroom.account.service.domain;
+
+public enum ClassroomAccountRole {
+    HOST,
+    GUEST
+}
