+++ src/main/java/soma/edupimeta/classroom/account/service/repository/ClassroomAccountRepository.java
@@ -1,8 +1,8 @@
-package soma.edupimeta.classroomAccount.service.repository;
+package soma.edupimeta.classroom.account.service.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import soma.edupimeta.classroomAccount.service.domain.ClassroomAccount;
+import soma.edupimeta.classroom.account.service.domain.ClassroomAccount;
 
 @Repository
 public interface ClassroomAccountRepository extends JpaRepository<ClassroomAccount, Long> {
