+++ src/main/java/soma/edupimeta/classroom/models/CreateClassroomRequest.java
@@ -10,12 +10,10 @@ public class CreateClassroomRequest {
 
     private final Long accountId;
     private final String name;
-    private final String inviteLink;
 
     public Classroom toEntity() {
         return Classroom.builder()
             .name(name)
-            .inviteLink(inviteLink)
             .build();
     }
