+++ src/main/java/soma/edupimeta/classroom/service/domain/Classroom.java
@@ -5,13 +5,16 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDateTime;
+import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
+import soma.edupimeta.classroom.account.service.domain.ClassroomAccount;
 
 @Entity
 @Getter
@@ -26,8 +29,8 @@ public class Classroom {
     @NotNull
     private String name;
 
-    @NotNull
-    private String inviteLink;
+    @OneToMany
+    private List<ClassroomAccount> classroomAccounts;
 
     @CreationTimestamp
     @Column(updatable = false)
@@ -37,8 +40,7 @@ public class Classroom {
     private LocalDateTime updatedAt;
 
     @Builder
-    public Classroom(String name, String inviteLink) {
+    public Classroom(String name) {
         this.name = name;
-        this.inviteLink = inviteLink;
     }
 }
