--- src/main/java/soma/edupimeta/classroomAccount/service/domain/ClassroomAccountRole.java
@@ -1,6 +1,0 @@
-package soma.edupimeta.classroomAccount.service.domain;
-
-public enum ClassroomAccountRole {
-    HOST,
-    GUEST;
-}
